Fix Gira Baskets
Heh, so, for plSynchedObject SDL updates, we were saving the update directly to the database instead of saving the merged blobs. This didn't work because the client only sends DIRTY fields in its updates. So, if the last basket update we get is an angular/velocity change with no position/orientation change, we've unset those fields in the DB. So, on subsequent link ins, they're at the DEFAULT, which is the origin (0, 0, 0). Winning!