Allow updating NULL SDL Nodes
Ages with no SDL are initialized with a null SDL blob. Those cannot be upgraded because there is no descriptor name in a null blob. This only happens for age SDL, so we can detect that nullness and see if a descriptor is actually available. Without this fix, the client will crash when linking to an age that has gained SDL.