Send entire files immediately.
Several Gehn Shard users have complained about download speeds from the
file server being limited to 300 KiB/s and lower, while I have been
seeing speeds of roughly 1 MiB/s. It appears that the RTT and waiting
for acks is extremely detrimental to download speeds. With this change,
I now see downloads at roughly 14 MiB/s, matching the rate of HTTPS
transfers.

The downside to this approach is that concurrent downloads are no longer
supported and no other request will be answered while a download is in
progress. This is not particularly concerning, however, because the
client requests files serially.