Fix crash triggered by clients sending too-new SDL.
If someone is hacking, developing, etc., then they may have an SDL
descriptor that is newer than what the server knows about. This is most
likely to happen for Age SDL (eg Personal, Kveer, etc). The client
blindly updates any out of date Age Vault SDL that it finds to whatever
it thinks the most recent descriptor is. This fixes the DS vault SDL
merging code to be more selective and not crash when receiving an
"invalid" SDL blob.

To reproduce, open the Personal.sdl file in your client, copy the last
STATEDESC, and increment its version.