Fix dockersand restart directives.
The previous values would cause the postgres container to restart when,
for example, the computer was restarted. The dirtsand container was left
unrestarted. We don't want to automatically restart the containers when
the computer boots up, I expect, and we definitely want for the
containers to have the same restart policy. It seems like the best
approach is to restart the containers when they crash only.