Better edge case handling around Docker Compose v2
The original version of Docker Compose was a separate utility, which
provided a docker-compose command-line tool.

Docker Compose v2 is a Docker plugin, which is meant to be run as a
subcommand of Docker (i.e., `docker compose ...`). In most cases, it
provides a `docker-compose` tool for backwards compatibility that
forwards to the plugin, but in some cases that tool is either missing or
not properly added to the path.

We will therefore try to run `docker-compose`, and if it fails we will
try to run `docker compose` and see if that works before exiting with an
error. If `docker compose` is supported, we will use that for subsequent
commands.