Send SDLs from the server SDL directory.
The client now prefers to download SDL files from the server, even when
`/LocalData` is supplied. Therefore, it is important that DS is always
able to serve SDL files without intervention. So, when SDLs are
requested from the AuthServ, send down an encrypted copy of the SDLs
that DS loaded.