Change Docker restart mode to unless-stopped
Previously this was `always` which caused containers to immediately start when the host computer booted, so it was changed to `on-failure` so that they'd restart if they crashed. There's still an edge case here where updating or restarting the docker daemon is considered failure and the containers still spin up on their own sometimes. I'm hoping that `unless-stopped` will be better about restarting on crashes but not in other circumstances.